Update/revert to apidef rpc model
