+++ internal/model/rpc.go
@@ -1,51 +1,34 @@
 package model
 
-type InboundData struct {
-	KeyName      string
-	Value        string
-	SessionState string
-	Timeout      int64
-	Per          int64
-	Expire       int64
-}
+import "github.com/TykTechnologies/tyk/apidef"
 
-type DefRequest struct {
-	OrgId   string
-	Tags    []string
-	LoadOAS bool
-}
+// This contains the shim for rpc data model types.
+// They are used from tests, and just pipe through
+// the apidef types to avoid import cycles.
+type (
+	GroupLoginRequest    = apidef.GroupLoginRequest
+	GroupKeySpaceRequest = apidef.GroupKeySpaceRequest
+	DefRequest           = apidef.DefRequest
+	InboundData          = apidef.InboundData
+	KeysValuesPair       = apidef.KeysValuesPair
+)
 
-type GroupLoginRequest struct {
-	UserKey   string
-	GroupID   string
-	ForceSync bool
-	Node      []byte
-}
+// These are health check shims.
+type (
+	HealthCheckItem     = apidef.HealthCheckItem
+	HealthCheckResponse = apidef.HealthCheckResponse
+	HealthCheckStatus   = apidef.HealthCheckStatus
 
-type NodeData struct {
-	NodeID          string                     `json:"node_id"`
-	APIKey          string                     `json:"api_key"`
-	GroupID         string                     `json:"group_id"`
-	NodeVersion     string                     `json:"node_version"`
-	TTL             int64                      `json:"ttl"`
-	NodeIsSegmented bool                       `json:"node_is_segmented"`
-	Tags            []string                   `json:"tags"`
-	Health          map[string]HealthCheckItem `json:"health"`
-	Stats           GWStats                    `json:"stats"`
-	HostDetails     HostDetails                `json:"host_details"`
-}
+	HostDetails = apidef.HostDetails
+	NodeData    = apidef.NodeData
+	GWStats     = apidef.GWStats
+)
 
-type GWStats struct {
-	APIsCount     int `json:"apis_count"`
-	PoliciesCount int `json:"policies_count"`
-}
-
-type GroupKeySpaceRequest struct {
-	OrgID   string
-	GroupID string
-}
-
-type KeysValuesPair struct {
-	Keys   []string
-	Values []string
-}
+// Other.
+const (
+	Pass      = apidef.Pass
+	Warn      = apidef.Warn
+	Fail      = apidef.Fail
+	System    = apidef.System
+	Datastore = apidef.Datastore
+)
