+++ rpc/synchronization_forcer_test.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"github.com/TykTechnologies/tyk/apidef"
 	"github.com/TykTechnologies/tyk/config"
-	"github.com/TykTechnologies/tyk/internal/model"
 	"github.com/TykTechnologies/tyk/storage"
 )
 
@@ -52,14 +52,14 @@ func TestGroupLoginCallback(t *testing.T) {
 	groupID := "group"
 
 	//first time, it should force since the group key doesn't exists
-	groupLogin, ok := sf.GroupLoginCallback(key, groupID).(model.GroupLoginRequest)
+	groupLogin, ok := sf.GroupLoginCallback(key, groupID).(apidef.GroupLoginRequest)
 	assert.True(t, ok)
 	assert.Equal(t, true, groupLogin.ForceSync)
 	assert.Equal(t, key, groupLogin.UserKey)
 	assert.Equal(t, groupID, groupLogin.GroupID)
 
 	//second time, it shouldn't force since the group key already exists
-	groupLogin, ok = sf.GroupLoginCallback(key, groupID).(model.GroupLoginRequest)
+	groupLogin, ok = sf.GroupLoginCallback(key, groupID).(apidef.GroupLoginRequest)
 	assert.True(t, ok)
 	assert.Equal(t, false, groupLogin.ForceSync)
 	assert.Equal(t, key, groupLogin.UserKey)
